reactive_graph_plugin_api::types::components::component_manager

Trait ComponentManager

pub trait ComponentManager: Send + Sync {

    // Required methods
    fn get_all(&self) -> Components;
    fn get_by_namespace(&self, namespace: &str) -> Components;
    fn has(&self, ty: &ComponentTypeId) -> bool;
    fn has_by_type(&self, namespace: &str, name: &str) -> bool;
    fn get(&self, ty: &ComponentTypeId) -> Option<Component>;
    fn get_by_type(&self, namespace: &str, name: &str) -> Option<Component>;
    fn find_by_type_name(&self, search: &str) -> Components;
    fn count(&self) -> usize;
    fn count_by_namespace(&self, namespace: &str) -> usize;
    fn create(
        &self,
        ty: &ComponentTypeId,
        description: &str,
        properties: PropertyTypes,
        extensions: Extensions,
    ) -> Result<Component, ComponentCreationError>;
    fn replace(&self, ty: &ComponentTypeId, component: Component);
    fn update_description(
        &self,
        ty: &ComponentTypeId,
        description: &str,
    ) -> Result<Component, ComponentUpdateError>;
    fn add_property(
        &self,
        ty: &ComponentTypeId,
        property: PropertyType,
    ) -> Result<PropertyType, ComponentAddPropertyError>;
    fn update_property(
        &self,
        ty: &ComponentTypeId,
        property_name: &str,
        property: PropertyType,
    ) -> Result<PropertyType, ComponentUpdatePropertyError>;
    fn remove_property(
        &self,
        ty: &ComponentTypeId,
        property_name: &str,
    ) -> Result<PropertyType, ComponentRemovePropertyError>;
    fn add_extension(
        &self,
        ty: &ComponentTypeId,
        extension: Extension,
    ) -> Result<ExtensionTypeId, ComponentAddExtensionError>;
    fn update_extension(
        &self,
        component_ty: &ComponentTypeId,
        extension_ty: &ExtensionTypeId,
        extension: Extension,
    ) -> Result<Extension, ComponentUpdateExtensionError>;
    fn remove_extension(
        &self,
        component_ty: &ComponentTypeId,
        extension_ty: &ExtensionTypeId,
    ) -> Result<Extension, ComponentRemoveExtensionError>;
    fn delete(&self, ty: &ComponentTypeId) -> bool;

}

Required Methods

fn get_all(&self) -> Components

Returns all components

fn get_by_namespace(&self, namespace: &str) -> Components

Returns all components of the given namespace.

fn has(&self, ty: &ComponentTypeId) -> bool

Returns true, if a component with the given name exists.

fn has_by_type(&self, namespace: &str, name: &str) -> bool

Returns true, if a component with the given fully qualified name exists.

fn get(&self, ty: &ComponentTypeId) -> Option<Component>

Returns the component with the given name or empty.

fn get_by_type(&self, namespace: &str, name: &str) -> Option<Component>

Returns the component with the given fully qualified name or empty.

fn find_by_type_name(&self, search: &str) -> Components

Returns all components whose names matches the given search string.

fn count(&self) -> usize

Returns the count of components.

fn count_by_namespace(&self, namespace: &str) -> usize

Returns the count of components of the given namespace.

fn create( &self, ty: &ComponentTypeId, description: &str, properties: PropertyTypes, extensions: Extensions, ) -> Result<Component, ComponentCreationError>

Creates a new component with the given name and the given properties.

fn replace(&self, ty: &ComponentTypeId, component: Component)

Replaces the component with the given name with the given component.

fn update_description( &self, ty: &ComponentTypeId, description: &str, ) -> Result<Component, ComponentUpdateError>

Updates the description of the given component.

fn add_property( &self, ty: &ComponentTypeId, property: PropertyType, ) -> Result<PropertyType, ComponentAddPropertyError>

Adds a property to the component with the given name.

fn update_property( &self, ty: &ComponentTypeId, property_name: &str, property: PropertyType, ) -> Result<PropertyType, ComponentUpdatePropertyError>

Adds a property to the component with the given name.

fn remove_property( &self, ty: &ComponentTypeId, property_name: &str, ) -> Result<PropertyType, ComponentRemovePropertyError>

Removes the property with the given property_name from the component with the given name.

fn add_extension( &self, ty: &ComponentTypeId, extension: Extension, ) -> Result<ExtensionTypeId, ComponentAddExtensionError>

Adds an extension to the component with the given name.

fn update_extension( &self, component_ty: &ComponentTypeId, extension_ty: &ExtensionTypeId, extension: Extension, ) -> Result<Extension, ComponentUpdateExtensionError>

Replaces the extension of the given component.

fn remove_extension( &self, component_ty: &ComponentTypeId, extension_ty: &ExtensionTypeId, ) -> Result<Extension, ComponentRemoveExtensionError>

Removes the extension with the given type from the component with the given type.

fn delete(&self, ty: &ComponentTypeId) -> bool

Deletes the component with the given name.