Trait reactive_graph_plugin_api::types::relations::relation_type_manager::RelationTypeManager

pub trait RelationTypeManager: Send + Sync {

    // Required methods
    fn get_all(&self) -> RelationTypes;
    fn get_by_namespace(&self, namespace: &str) -> RelationTypes;
    fn has(&self, ty: &RelationTypeId) -> bool;
    fn has_by_type(&self, namespace: &str, type_name: &str) -> bool;
    fn get(&self, ty: &RelationTypeId) -> Option<RelationType>;
    fn get_by_type(
        &self,
        namespace: &str,
        type_name: &str,
    ) -> Option<RelationType>;
    fn find_by_type_name(&self, search: &str) -> RelationTypes;
    fn count(&self) -> usize;
    fn count_by_namespace(&self, namespace: &str) -> usize;
    fn create(
        &self,
        outbound_type: &ComponentOrEntityTypeId,
        ty: &RelationTypeId,
        inbound_type: &ComponentOrEntityTypeId,
        description: &str,
        components: ComponentTypeIds,
        properties: PropertyTypes,
        extensions: Extensions,
    ) -> Result<RelationType, RelationTypeCreationError>;
    fn update_description(
        &self,
        ty: &RelationTypeId,
        description: &str,
    ) -> Result<RelationType, RelationTypeUpdateError>;
    fn add_component(
        &self,
        ty: &RelationTypeId,
        component: &ComponentTypeId,
    ) -> Result<(), RelationTypeAddComponentError>;
    fn remove_component(
        &self,
        ty: &RelationTypeId,
        component: &ComponentTypeId,
    ) -> Result<ComponentTypeId, RelationTypeRemoveComponentError>;
    fn add_property(
        &self,
        ty: &RelationTypeId,
        property: PropertyType,
    ) -> Result<PropertyType, RelationTypeAddPropertyError>;
    fn update_property(
        &self,
        relation_ty: &RelationTypeId,
        property_name: &str,
        property_type: PropertyType,
    ) -> Result<PropertyType, RelationTypeUpdatePropertyError>;
    fn remove_property(
        &self,
        ty: &RelationTypeId,
        property_name: &str,
    ) -> Result<PropertyType, RelationTypeRemovePropertyError>;
    fn add_extension(
        &self,
        ty: &RelationTypeId,
        extension: Extension,
    ) -> Result<ExtensionTypeId, RelationTypeAddExtensionError>;
    fn update_extension(
        &self,
        relation_ty: &RelationTypeId,
        extension_ty: &ExtensionTypeId,
        extension: Extension,
    ) -> Result<Extension, RelationTypeUpdateExtensionError>;
    fn remove_extension(
        &self,
        relation_ty: &RelationTypeId,
        extension_ty: &ExtensionTypeId,
    ) -> Result<Extension, RelationTypeRemoveExtensionError>;
    fn delete(&self, ty: &RelationTypeId) -> Option<RelationType>;
    fn validate(&self, ty: &RelationTypeId) -> bool;

}

Required Methods

fn get_all(&self) -> RelationTypes

Returns all relation types.

fn get_by_namespace(&self, namespace: &str) -> RelationTypes

Returns all relation types of the given namespace.

fn has(&self, ty: &RelationTypeId) -> bool

Returns true, if a relation type with the given name exists.

fn has_by_type(&self, namespace: &str, type_name: &str) -> bool

Returns true, if a relation type with the given fully qualified name exists.

fn get(&self, ty: &RelationTypeId) -> Option<RelationType>

Returns the relation type with the given name.

fn get_by_type(&self, namespace: &str, type_name: &str) -> Option<RelationType>

Returns the relation type with the given fully qualified name.

fn find_by_type_name(&self, search: &str) -> RelationTypes

Returns all relation types whose names matches the given search string.

fn count(&self) -> usize

Returns the count of relation types.

fn count_by_namespace(&self, namespace: &str) -> usize

Returns the count of relation types of the given namespace.

fn create( &self, outbound_type: &ComponentOrEntityTypeId, ty: &RelationTypeId, inbound_type: &ComponentOrEntityTypeId, description: &str, components: ComponentTypeIds, properties: PropertyTypes, extensions: Extensions, ) -> Result<RelationType, RelationTypeCreationError>

Creates a new relation type.

fn update_description( &self, ty: &RelationTypeId, description: &str, ) -> Result<RelationType, RelationTypeUpdateError>

Updates the description of the given relation type.

fn add_component( &self, ty: &RelationTypeId, component: &ComponentTypeId, ) -> Result<(), RelationTypeAddComponentError>

Adds the component with the given type to the given relation type.

fn remove_component( &self, ty: &RelationTypeId, component: &ComponentTypeId, ) -> Result<ComponentTypeId, RelationTypeRemoveComponentError>

Remove the component with the given type from the given relation type.

fn add_property( &self, ty: &RelationTypeId, property: PropertyType, ) -> Result<PropertyType, RelationTypeAddPropertyError>

Adds a property to the given relation type.

fn update_property( &self, relation_ty: &RelationTypeId, property_name: &str, property_type: PropertyType, ) -> Result<PropertyType, RelationTypeUpdatePropertyError>

Updates the property with the given property_name. It’s possible to rename the property by using another name in the new property than the provided property_name.

fn remove_property( &self, ty: &RelationTypeId, property_name: &str, ) -> Result<PropertyType, RelationTypeRemovePropertyError>

Removes the property with the given property_name from the given relation type.

fn add_extension( &self, ty: &RelationTypeId, extension: Extension, ) -> Result<ExtensionTypeId, RelationTypeAddExtensionError>

Adds an extension to the given relation type.

fn update_extension( &self, relation_ty: &RelationTypeId, extension_ty: &ExtensionTypeId, extension: Extension, ) -> Result<Extension, RelationTypeUpdateExtensionError>

fn remove_extension( &self, relation_ty: &RelationTypeId, extension_ty: &ExtensionTypeId, ) -> Result<Extension, RelationTypeRemoveExtensionError>

Removes the extension with the given type from the given relation type.

fn delete(&self, ty: &RelationTypeId) -> Option<RelationType>

Deletes the given relation type.

fn validate(&self, ty: &RelationTypeId) -> bool

Validates the relation type with the given name. Tests that all components, the outbound and inbound entity type exists.