Trait reactive_graph_plugin_api::prelude::plugin::TypeProvider

pub trait TypeProvider<T>: Send + Sync
where
    T: NamespacedTypeContainer,
{
    // Required methods
    fn id<'a>(&self) -> &'a str;
    fn get_types(&self) -> T;

    // Provided method
    fn get_type_ids(&self) -> <T as NamespacedTypeContainer>::TypeIds { ... }
}

Required Methods

fn id<'a>(&self) -> &'a str

Returns the id of the type provider.

fn get_types(&self) -> T

Returns a collection of types which should be registered.

Provided Methods

fn get_type_ids(&self) -> <T as NamespacedTypeContainer>::TypeIds

Returns a collection of type ids which should be unregistered.