Trait reactive_graph_plugin_api::types::flows::flow_type_manager::FlowTypeManager

pub trait FlowTypeManager: Send + Sync {

    // Required methods
    fn get_all(&self) -> FlowTypes;
    fn get_by_namespace(&self, namespace: &str) -> FlowTypes;
    fn has(&self, ty: &FlowTypeId) -> bool;
    fn has_by_type(&self, namespace: &str, name: &str) -> bool;
    fn get(&self, ty: &FlowTypeId) -> Option<FlowType>;
    fn get_by_type(&self, namespace: &str, name: &str) -> Option<FlowType>;
    fn find_by_type_name(&self, search: &str) -> FlowTypes;
    fn count(&self) -> usize;
    fn count_by_namespace(&self, namespace: &str) -> usize;
    fn create(
        &self,
        ty: &FlowTypeId,
        description: &str,
        wrapper_entity_instance: EntityInstance,
        entity_instances: EntityInstances,
        relation_instances: RelationInstances,
        variables: PropertyTypes,
        extensions: Extensions,
    ) -> Result<FlowType, FlowTypeCreationError>;
    fn add_entity_instance(
        &self,
        ty: &FlowTypeId,
        entity_instance: EntityInstance,
    ) -> Result<(), FlowTypeAddEntityInstanceError>;
    fn update_entity_instance(
        &self,
        ty: &FlowTypeId,
        id: Uuid,
        entity_instance: EntityInstance,
    ) -> Result<(Uuid, EntityInstance), FlowTypeUpdateEntityInstanceError>;
    fn remove_entity_instance(
        &self,
        ty: &FlowTypeId,
        id: Uuid,
    ) -> Result<Option<(Uuid, EntityInstance)>, FlowTypeRemoveEntityInstanceError>;
    fn add_extension(
        &self,
        ty: &FlowTypeId,
        extension: Extension,
    ) -> Result<ExtensionTypeId, FlowTypeAddExtensionError>;
    fn update_extension(
        &self,
        ty: &FlowTypeId,
        extension_ty: &ExtensionTypeId,
        extension: Extension,
    ) -> Result<Extension, FlowTypeUpdateExtensionError>;
    fn remove_extension(
        &self,
        flow_ty: &FlowTypeId,
        extension_ty: &ExtensionTypeId,
    ) -> Result<Extension, FlowTypeRemoveExtensionError>;
    fn add_variable(
        &self,
        ty: &FlowTypeId,
        variable: PropertyType,
    ) -> Result<Variable, FlowTypeAddVariableError>;
    fn update_variable(
        &self,
        ty: &FlowTypeId,
        variable_name: &str,
        variable: PropertyType,
    ) -> Result<Variable, FlowTypeUpdateVariableError>;
    fn remove_variable(
        &self,
        ty: &FlowTypeId,
        variable_name: &str,
    ) -> Result<Variable, FlowTypeRemoveVariableError>;
    fn delete(&self, ty: &FlowTypeId) -> Option<FlowType>;
    fn validate(&self, ty: &FlowTypeId) -> bool;

}

Required Methods

fn get_all(&self) -> FlowTypes

Returns all flow types.

fn get_by_namespace(&self, namespace: &str) -> FlowTypes

Returns all flow types.

fn has(&self, ty: &FlowTypeId) -> bool

Returns true, if a flow type with the given name exists.

fn has_by_type(&self, namespace: &str, name: &str) -> bool

Returns true, if a flow type with the given name exists.

fn get(&self, ty: &FlowTypeId) -> Option<FlowType>

Returns the flow type with the given name or empty.

fn get_by_type(&self, namespace: &str, name: &str) -> Option<FlowType>

Returns the flow type with the given name or empty.

fn find_by_type_name(&self, search: &str) -> FlowTypes

Returns all flow types whose names matches the given search string.

fn count(&self) -> usize

Returns the count of flow types.

fn count_by_namespace(&self, namespace: &str) -> usize

Returns the count of flow types of the given namespace.

fn create( &self, ty: &FlowTypeId, description: &str, wrapper_entity_instance: EntityInstance, entity_instances: EntityInstances, relation_instances: RelationInstances, variables: PropertyTypes, extensions: Extensions, ) -> Result<FlowType, FlowTypeCreationError>

Creates a new flow type.

fn add_entity_instance( &self, ty: &FlowTypeId, entity_instance: EntityInstance, ) -> Result<(), FlowTypeAddEntityInstanceError>

Adds the given entity instance to the flow type with the given name.

fn update_entity_instance( &self, ty: &FlowTypeId, id: Uuid, entity_instance: EntityInstance, ) -> Result<(Uuid, EntityInstance), FlowTypeUpdateEntityInstanceError>

Updates the entity instance with the given id of the flow type with the given name.

fn remove_entity_instance( &self, ty: &FlowTypeId, id: Uuid, ) -> Result<Option<(Uuid, EntityInstance)>, FlowTypeRemoveEntityInstanceError>

Removes the entity instance with the given id from the flow type with the given name.

fn add_extension( &self, ty: &FlowTypeId, extension: Extension, ) -> Result<ExtensionTypeId, FlowTypeAddExtensionError>

Adds the given extension to the given flow type.

fn update_extension( &self, ty: &FlowTypeId, extension_ty: &ExtensionTypeId, extension: Extension, ) -> Result<Extension, FlowTypeUpdateExtensionError>

Updates the extension with the given type of the given flow type.

fn remove_extension( &self, flow_ty: &FlowTypeId, extension_ty: &ExtensionTypeId, ) -> Result<Extension, FlowTypeRemoveExtensionError>

Removes the extension with the given type from the given flow type.

fn add_variable( &self, ty: &FlowTypeId, variable: PropertyType, ) -> Result<Variable, FlowTypeAddVariableError>

Adds the given variable to the given flow type.

fn update_variable( &self, ty: &FlowTypeId, variable_name: &str, variable: PropertyType, ) -> Result<Variable, FlowTypeUpdateVariableError>

Updates the variable with the given name of the flow type with the given name.

fn remove_variable( &self, ty: &FlowTypeId, variable_name: &str, ) -> Result<Variable, FlowTypeRemoveVariableError>

Removes the variable with the given name from the flow type with the given name.

fn delete(&self, ty: &FlowTypeId) -> Option<FlowType>

Deletes the flow type with the given name.

fn validate(&self, ty: &FlowTypeId) -> bool

Validates the flow type with the given name. Tests that all entity types and relation types exists and are valid.